# Conflicts:
#	app/src/main/java/com/aayaffe/sailingracecoursemanager/Calc_Layer/Mark.java

+++ app/src/main/java/com/aayaffe/sailingracecoursemanager/Calc_Layer/RaceCourse.java
@@ -22,7 +22,7 @@ public class RaceCourse {
     private static double dist2m1;  //distance to mark 1
     private static int windDir;  //wind direction
     private static AviLocation signalBoatLoc;
-    private static float startLineDist;
+    private static double startLineDist;
     private static Map<String, String> selectedOptions;
     private static List<Buoy> bouyList = new ArrayList<Buoy>();
     public static CourseXmlParser xmlParserC;
@@ -32,7 +32,7 @@ public class RaceCourse {
         //Empty constructor for Serializing to firebase
     }
 
-    public RaceCourse(Context context, AviLocation signalBoatLoc, int windDirection, float distance2mark1 , float startLineDistance,Map<String, String> selectedCourseOptions ){
+    public RaceCourse(Context context, AviLocation signalBoatLoc, int windDirection, double distance2mark1 , double startLineDistance,Map<String, String> selectedCourseOptions ){
         this.context = context;
         if(signalBoatLoc!=null) RaceCourse.signalBoatLoc =signalBoatLoc;
         else RaceCourse.signalBoatLoc =new AviLocation(32.85,3499);
